# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin-mall API Documentation",
        "title": "gin-mall",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "1478488313@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "输入`用户名-密码`以登录",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.AuthLoginResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "输入`用户名-密码-确认密码`以注册",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "确认密码",
                        "name": "confirm_password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.AuthRegisterResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/product/{id}": {
            "get": {
                "description": "通过商品ID获取商品详情信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "获取商品详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "商品id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.GetProductListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "通过页号和页码获取指定的商品列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "获取商品列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页号",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "大小",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.GetProductListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/e.Code"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "e.Code": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15,
                16,
                17,
                18,
                19,
                20,
                21,
                22,
                23,
                24,
                25,
                26,
                27,
                28,
                29,
                30
            ],
            "x-enum-comments": {
                "ErrorAccountInvalid": "用户名或密码错误",
                "ErrorContextValue": "上下文值传递错误",
                "ErrorCreateUser": "创建用户错误",
                "ErrorEmailLinkExpire": "邮件确认链接已过期",
                "ErrorEncryptMoney": "金额加密错误",
                "ErrorEncryptPassword": "密码加密错误",
                "ErrorFileError": "文件错误",
                "ErrorFileType": "文件类型错误",
                "ErrorFollowUser": "关注用户失败",
                "ErrorGenerateToken": "token生成错误",
                "ErrorGetCategoryList": "获取商品分类失败",
                "ErrorGetFollowerList": "获取粉丝列表失败",
                "ErrorGetFollowingList": "获取关注列表失败",
                "ErrorGetProductByID": "根据ID获取商品失败",
                "ErrorGetProductList": "获取商品列表失败",
                "ErrorGetUserByID": "根据id获取用户失败",
                "ErrorIncorrectPassword": "密码错误",
                "ErrorInvalidIDParam": "非法的id参数",
                "ErrorOSSUploadError": "OSS文件上传错误",
                "ErrorParseToken": "token解析错误",
                "ErrorSendEmail": "发送邮件错误",
                "ErrorSendEmailTooFrequent": "邮件发送操作频繁",
                "ErrorUnfollowUser": "取消关注失败",
                "ErrorUpdateEmail": "更新邮箱错误",
                "ErrorUpdateUser": "更新用户失败",
                "ErrorUploadAvatar": "头像上传错误",
                "ErrorUploadFile": "文件上传错误",
                "ErrorUserExists": "用户已存在",
                "InvalidParams": "参数错误",
                "Success": "响应成功",
                "UnknownError": "未知错误"
            },
            "x-enum-varnames": [
                "Success",
                "InvalidParams",
                "UnknownError",
                "ErrorUserExists",
                "ErrorEncryptPassword",
                "ErrorEncryptMoney",
                "ErrorCreateUser",
                "ErrorAccountInvalid",
                "ErrorGetUserByID",
                "ErrorUpdateUser",
                "ErrorIncorrectPassword",
                "ErrorUploadAvatar",
                "ErrorFollowUser",
                "ErrorUnfollowUser",
                "ErrorGetFollowingList",
                "ErrorGetFollowerList",
                "ErrorGenerateToken",
                "ErrorParseToken",
                "ErrorContextValue",
                "ErrorUploadFile",
                "ErrorFileError",
                "ErrorOSSUploadError",
                "ErrorFileType",
                "ErrorSendEmail",
                "ErrorUpdateEmail",
                "ErrorEmailLinkExpire",
                "ErrorSendEmailTooFrequent",
                "ErrorGetCategoryList",
                "ErrorGetProductList",
                "ErrorGetProductByID",
                "ErrorInvalidIDParam"
            ]
        },
        "response.AuthLoginResp": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "response.AuthRegisterResp": {
            "type": "object"
        },
...
        "response.GetProductListResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Product"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.Product": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "商品ID",
                    "type": "integer"
                },
                "image_url": {
                    "description": "商品主图的URL",
                    "type": "string"
                },
                "price": {
                    "description": "商品价格",
                    "type": "number"
                },
                "stock": {
                    "description": "库存数量",
                    "type": "integer"
                },
                "title": {
                    "description": "商品名称",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "RefreshToken": {
            "type": "apiKey",
            "name": "X-Refresh-Token",
            "in": "header"
        }
    },
...
}